fix(libs): checked untrusted lengths, header-field overflow errors, no stranded .tmp
cpk, ftex and pes_savefile narrowed file-declared u64 lengths and multiplied
header dimensions unchecked (wrong on a 32-bit usize, where pes_model already
did it right); FTEX header fields took DDS dimensions with a wrapping cast
(HeaderFieldOverflow now). Settings::save and EditFile::save remove their
.tmp sibling when the write or rename fails. dds_convert no longer clones
every mip level in generate() or on the non-normal encode path. pes_savefile
crate doc describes what exists. Small doc fixes in color_tools, examples/skl.py
and a valid .vscode/settings.json.

+++ crates/libs/ftex/src/format.rs
@@ -177,7 +177,8 @@ pub fn mip_size(format: PixelFormat, width: u32, height: u32, depth: u32, level:
     let d = (depth >> level).max(1);
     let blocks_w = w.div_ceil(block_pixels);
     let blocks_h = h.div_ceil(block_pixels);
-    (blocks_w * blocks_h * d * block_bytes) as usize
+    // u64: the u32 product wraps for a 64k x 64k x deep-enough volume.
+    (u64::from(blocks_w) * u64::from(blocks_h) * u64::from(d) * u64::from(block_bytes)) as usize
 }
 
 /// The 64-byte FTEX header (little-endian).
